able to extend from fixedAsset and fixedAsset implement Valuable

+++ workbook-5/exercises/MajesticYardCare/src/main/java/com/pluralsight/models/Circle.java
@@ -16,7 +16,7 @@ public class Circle extends Shape
     public double getArea()
     {
 
-        return PI * Math.pow(radius,2);
+        return PI * (int) Math.pow(radius,2);
     }
 
     @Override
